shyam_sunder_reddy changes in tasks patch and update based on role

+++ shyam_sunder_reddy/ust_employee_task_management/backend/routers/task_api.py
@@ -1,5 +1,5 @@
 from fastapi import APIRouter, HTTPException,Depends
-from crud.task_crud import add_task, get_all_tasks, get_task_by_id, update_task, delete_task
+from crud.task_crud import add_task, get_all_tasks, get_task_by_id, update_task, delete_task,get_task_by_status,patch_status
 from models.task import TaskReqRes
 from typing import List
 from utils.auth import get_current_user
@@ -44,10 +44,26 @@ def get_by_id(id: int,user=Depends(get_current_user)):
         raise HTTPException(status_code=500, detail=f"Internal Server Error: {str(e)}")
 
 
-@task_router.put("/update")
-def update_task_data(id: int, new_data: dict,user=Depends(get_current_user)):
+@task_router.get("/getbystatus",response_model=List[TaskReqRes])
+def get_by_status(status,role,user=Depends(get_current_user)):
     try:
-        updated = update_task(id, new_data,user)
+        if role not in user.role:
+            raise HTTPException(status_code=409,detail="The user doesnt have the mentioned role")
+        return get_task_by_status(status,role,user)
+    except HTTPException as e:
+        raise e
+    except Exception as e:
+        raise HTTPException(status_code=500, detail=f"Internal Server Error: {str(e)}")
+
+
+@task_router.put("/update")
+def update_task_data(id: int, new_data: dict,role,user=Depends(get_current_user)):
+    try:
+        if role not in user.role:
+            raise HTTPException(status_code=409,detail="The user doesnt have the mentioned role")
+        if role=="Admin" and new_data["status"]:
+            raise HTTPException(status_code=409,detail="Admin cannout update the status of task")    
+        updated = update_task(id,new_data,role,user)
         return {"detail": "Task Updated Successfully", "task": updated}
     except HTTPException as e:
         raise e
@@ -55,9 +71,24 @@ def update_task_data(id: int, new_data: dict,user=Depends(get_current_user)):
         raise HTTPException(status_code=500, detail=f"Internal Server Error: {str(e)}")
 
 
-@task_router.delete("/delete")
-def delete_task_by_id(id: int,user=Depends(get_current_user)):
+@task_router.patch("/patch")
+def patch_stat(id,status,role,user=Depends(get_current_user)):
     try:
+        if role not in user.role and role.upper() !="ADMIN":
+            raise HTTPException(status_code=409,detail="The user doesnt have the mentioned role")
+        patched=patch_status(id,status,role,user)
+        return {"detail":"Patched the task"}
+    except HTTPException as e:
+        raise e
+    except Exception as e:
+        raise HTTPException(status_code=500, detail=f"Internal Server Error: {str(e)}")
+
+
+@task_router.delete("/delete")
+def delete_task_by_id(id: int,role,user=Depends(get_current_user)):
+    try:
+        if role.upper()!="ADMIN" and role not in user.role:
+            raise HTTPException(status_code=409,detail="Only the Admincan delete the task")
         resp = delete_task(id,user)
         return resp
     except HTTPException as e:
